+++ library/src/main/jni/com_twilio_video_ConnectOptions.h
@@ -40,6 +40,7 @@ JNIEXPORT jlong JNICALL Java_com_twilio_video_ConnectOptions_nativeCreate
          jlong,
          jobjectArray,
          jobjectArray,
+         jstring,
          jobject);